+++ examples/SSD1306_HELLO/main.cpp
@@ -10,15 +10,22 @@
 */
 
 // === Libraries ===
-#include <stdio.h>
+#include <cstdio>
 #include "pico/stdlib.h"
 #include "ssd1306/SSD1306_OLED.hpp"
 
+// Screen settings
 #define myOLEDwidth  128
 #define myOLEDheight 64
-const uint16_t I2C_Speed = 100; 
+#define myScreenSize (myOLEDwidth * (myOLEDheight/8)) // eg 1024 bytes = 128 * 64/8
+uint8_t screenBuffer[myScreenSize]; // Define a buffer to cover whole screen  128 * 64/8
+
+// I2C settings
+const uint16_t I2C_Speed = 100;
 const uint8_t I2C_Address = 0x3C;
-SSD1306 myOLED(myOLEDwidth ,myOLEDheight) ; // instantiate  an object 
+
+// instantiate  an OLED object
+SSD1306 myOLED(myOLEDwidth ,myOLEDheight);
 
 // =============== Function prototype ================
 void SetupTest(void);
@@ -38,27 +45,28 @@ int main()
 void SetupTest() 
 {
 	stdio_init_all(); // Initialize chosen serial port, default 38400 baud
-	busy_wait_ms(100);
+	busy_wait_ms(500);
 	printf("OLED SSD1306 :: Start!\r\n");
 	while(myOLED.OLEDbegin(I2C_Address,i2c1,  I2C_Speed, 18, 19) != true)
 	{
-		printf("OLED SSD1306 :: Failed to initialize OLED.!\r\n");
+		printf("SetupTest ERROR : Failed to initialize OLED!\r\n");
 		busy_wait_ms(1500);
 	} // initialize the OLED
+	if (myOLED.OLEDSetBufferPtr(myOLEDwidth, myOLEDheight, screenBuffer, sizeof(screenBuffer)/sizeof(uint8_t)) != 0)
+	{
+		printf("SetupTest : ERROR : OLEDSetBufferPtr Failed!\r\n");
+		while(1){busy_wait_ms(1000);}
+	} // Initialize the buffer
 	myOLED.OLEDFillScreen(0xF0, 0); // splash screen bars
 	busy_wait_ms(1000);
 }
 
 void Test() 
 {
-	// Define a buffer to cover whole screen 
-	uint8_t  screenBuffer[myOLEDwidth * (myOLEDheight/8)+1]; 
- 	myOLED.buffer = (uint8_t*) &screenBuffer;  // set that to library buffer pointer
 	myOLED.OLEDclearBuffer(); 
-	
-	myOLED.setFont(pFontDefault); 
-	myOLED.setCursor(5, 5);
-	myOLED.print("HELLO World!");
+	myOLED.setFont(pFontDefault);
+	myOLED.setCursor(5,5);
+	myOLED.print("Hello World!");
 	myOLED.OLEDupdate();  
 	busy_wait_ms(7000);
 }