+++ src/ssd1306/SSD1306_OLED.cpp
@@ -5,7 +5,7 @@
 * @details <https://github.com/gavinlyonsrepo/SSD1306_OLED_PICO>
 */
 
-#include <stdio.h> // optional for printf debug messages
+//#include <stdio.h> 
 #include "pico/stdlib.h"
 #include "../include/ssd1306/SSD1306_OLED.hpp"
 
@@ -19,15 +19,15 @@ SSD1306  :: SSD1306(int16_t oledwidth, int16_t oledheight) :SSD1306_graphics(ole
 	_OLED_HEIGHT = oledheight;
 	_OLED_WIDTH = oledwidth;
 	_OLED_PAGE_NUM = (_OLED_HEIGHT/8); 
-	bufferWidth = _OLED_WIDTH;
-	bufferHeight = _OLED_HEIGHT;
+	_bufferWidth = _OLED_WIDTH;
+	_bufferHeight = _OLED_HEIGHT;
 }
 
 /*!
 	@brief  begin Method initialise OLED I2C communication
 	@param I2Caddress I2C Bus address by default 0x3C
 	@param  i2c_type The I2C interface i2c0 or ic21 interface
-	@param CLKspeed I2C Bus Clock speed in KHz. Set to 100
+	@param CLKspeed I2C Bus Clock speed in KHz. 
 	@param SDApin I2C data GPIO pin  
 	@param SCLKpin I2C clock GPIO pin 
 	@return true if successful , init I2C communication
@@ -51,7 +51,7 @@ bool SSD1306::OLEDbegin( uint8_t I2Caddress, i2c_inst_t* i2c_type, uint16_t CLKs
 	i2c_init(_i2c, _CLKSpeed * 1000);
 
 	// check connection?
-	ReturnCode = i2c_read_timeout_us(_i2c, _OLEDAddressI2C , &rxData, 1, false, _OLEDdelayI2C);
+	ReturnCode = i2c_read_timeout_us(_i2c, _OLEDAddressI2C , &rxData, 1, false, _TimeoutDelayI2C);
 	if (ReturnCode < 1){ // no bytes read back from device or error issued
 		if (_bSerialDebugFlag)
 		{
@@ -66,6 +66,32 @@ bool SSD1306::OLEDbegin( uint8_t I2Caddress, i2c_inst_t* i2c_type, uint16_t CLKs
 	return true;
 }
 
+/*!
+	@brief sets the buffer pointer to the users screen data buffer
+	@param width width of buffer in pixels
+	@param height height of buffer in pixels
+	@param pBuffer the buffer array which decays to pointer
+	@param sizeOfBuffer size of buffer
+	@return Will return:
+		-# 0. Success
+		-# 2. Buffer size calculations are incorrect BufferSize = w * (h/8),
+		-# 3. Not a valid pointer object.
+*/
+uint8_t SSD1306::OLEDSetBufferPtr(uint8_t width, uint8_t height , uint8_t* pBuffer, uint16_t sizeOfBuffer)
+{
+	if(sizeOfBuffer !=  width * (height/8))
+	{
+		printf("SSD1306::OLEDSetBufferPtr Error 2: buffer size does not equal: w * (h/8))\r\n");
+		return 2;
+	}
+	_OLEDbuffer = pBuffer;
+	if(_OLEDbuffer ==  nullptr)
+	{
+		printf("SSD1306::OLEDSetBufferPtr Error 3: Problem assigning buffer pointer\r\n");
+		return 3;
+	}
+	return 0;
+}
 
 /*!
 	@brief End I2C operations. I2C pins P1-03 (SDA) and P1-05 (SCL) 	
@@ -217,12 +243,46 @@ void SSD1306::OLEDFillPage(uint8_t page_num, uint8_t dataPattern,uint8_t mydelay
 	@param y y axis offset 0-64
 	@param w width 0-128
 	@param h height 0-64
-	@param data pointer to bitmap data
+	@param pBitmap pointer to bitmap data
 	@param invert color 
-	@note bitmap data must be  horizontally addressed.
+	@param sizeOfBitmap size of bitmap
+	@return Will return 
+		-# 0. success
+		-# 2. Bitmap co-ord out of bounds, check x and y
+		-# 3. Bitmap is larger than screen, check w and h
+		-# 4. Bitmap is is not valid pointer
+		-# 5. Check Horizontal bitmap size
+	@note bitmap data must be  horizontally addressed and height divisible by 8
 */
-void SSD1306::OLEDBitmap(int16_t x, int16_t y, int16_t w, int16_t h, const uint8_t* data, bool invert)
+uint8_t SSD1306::OLEDBitmap(int16_t x, int16_t y, int16_t w, int16_t h, const uint8_t* pBitmap, bool invert,uint16_t sizeOfBitmap)
 {
+
+// User error checks
+// 1. Completely out of bounds?
+if (x > _width || y > _height)
+{
+	printf("SSD1306::OLEDBitmap Error 2: Bitmap co-ord out of bounds, check x and y\r\n");
+	return 2;
+}
+// 2. bitmap weight and height
+if (w > _width || h > _height)
+{
+	printf("SSD1306::OLEDBitmap Error 3: Bitmap is larger than screen, check w and h\r\n");
+	return 3;
+}
+// 3. bitmap is null
+if(pBitmap == nullptr)
+{
+	printf("SSD1306::OLEDBitmap Error 4: Bitmap is is not valid pointer\n");
+	return 4;
+}
+// 4. check Horizontal bitmap size
+if(sizeOfBitmap != ((w/8) * h))
+{
+	printf("SSD1306::OLEDBitmap Error 5: Horizontal Bitmap size is incorrect: %u  w %i h %i \n", sizeOfBitmap , w , h);
+	printf("Check size = ((w/8)*h) or is bitmap width divisible evenly by eight or is all bitmap data there or too much \n");
+	return 5;
+}
 
 int16_t byteWidth = (w + 7) / 8; 
 uint8_t byte = 0;
@@ -244,12 +304,12 @@ for (int16_t j = 0; j < h; j++, y++)
 		if (i & 7)
 			byte <<= 1;
 		else
-			byte = data[j * byteWidth + i / 8];
+			byte = pBitmap[j * byteWidth + i / 8];
 			
 		drawPixel(x + i, y, (byte & 0x80) ? color : bgcolor );
 	}
 }
-
+return 0;
 }
 
 /*!
@@ -262,24 +322,27 @@ void SSD1306::I2C_Write_Byte(uint8_t value, uint8_t cmd)
 {
 	uint8_t  dataBuffer[2] = {cmd,value};
 	uint8_t attemptI2Cwrite = 0;
-	uint8_t returnCode = 0;
+	int16_t returnCode = 0;
 	
 	//returnCode = bcm2835_i2c_write(buf, 2); 
-	returnCode = i2c_write_timeout_us(_i2c, _OLEDAddressI2C, dataBuffer, 2 , false, _OLEDdelayI2C);
+	returnCode = i2c_write_timeout_us(_i2c, _OLEDAddressI2C, dataBuffer, 2 , false, _TimeoutDelayI2C);
 	
 	while(returnCode < 1)
 	{ // failure to write I2C byte 
-		attemptI2Cwrite ++;
 		if (_bSerialDebugFlag)
 		{
-			printf("Error 1202 I2C OLED: Cannot Write byte :: %u\n", attemptI2Cwrite);
-			printf("Error code %u\n", returnCode);
+			printf("SSD1306::I2C_Write_Byte : Cannot Write byte : Retry Attempt = %u\n", attemptI2Cwrite);
+			printf("Error code %i\n", returnCode);
 		}
-		returnCode = i2c_write_timeout_us(_i2c, _OLEDAddressI2C, dataBuffer, 2 , false, _OLEDdelayI2C);
-		busy_wait_ms(100); // mS
-		if (attemptI2Cwrite >= 3) break;
+		if (attemptI2Cwrite >= _I2CRetryAttempts) break;
+		returnCode = i2c_write_timeout_us(_i2c, _OLEDAddressI2C, dataBuffer, 2 , false, _TimeoutDelayI2C);
+		busy_wait_ms(_I2CRetryDelay); // mS
+		attemptI2Cwrite ++;
 	}
-	if (returnCode < 1) _bIsConnected = false;
+	if (returnCode < 1) 
+		_bIsConnected = false;
+	else
+		_bIsConnected = true;
 }
 
 /*!
@@ -287,8 +350,8 @@ void SSD1306::I2C_Write_Byte(uint8_t value, uint8_t cmd)
 */
 void SSD1306::OLEDupdate()
 {
-	uint8_t x = 0; uint8_t y = 0; uint8_t w = this->bufferWidth; uint8_t h = this->bufferHeight;
-	OLEDBuffer( x,  y,  w,  h, (uint8_t*) this->buffer);
+	uint8_t x = 0; uint8_t y = 0; uint8_t w = this->_bufferWidth; uint8_t h = this->_bufferHeight;
+	OLEDBuffer( x,  y,  w,  h, (uint8_t*) this->_OLEDbuffer);
 }
 
 /*!
@@ -296,7 +359,7 @@ void SSD1306::OLEDupdate()
 */
 void SSD1306::OLEDclearBuffer()
 {
-	memset( this->buffer, 0x00, (this->bufferWidth * (this->bufferHeight /8))  );
+	memset( this->_OLEDbuffer, 0x00, (this->_bufferWidth * (this->_bufferHeight /8))  );
 }
 
 /*!
@@ -350,7 +413,7 @@ void SSD1306::OLEDBuffer(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t* da
 void SSD1306::drawPixel(int16_t x, int16_t y, uint8_t color)
 {
 
-	if ((x < 0) || (x >= this->bufferWidth) || (y < 0) || (y >= this->bufferHeight)) {
+	if ((x < 0) || (x >= this->_bufferWidth) || (y < 0) || (y >= this->_bufferHeight)) {
 	return;
 	}
 	int16_t temp;
@@ -370,12 +433,12 @@ void SSD1306::drawPixel(int16_t x, int16_t y, uint8_t color)
 		y = HEIGHT - 1 - temp;
 	break;
 	}
-		uint16_t tc = (bufferWidth * (y /8)) + x;
+		uint16_t tc = (_bufferWidth * (y /8)) + x;
 		switch (color)
 		{
-			case WHITE:  this->buffer[tc] |= (1 << (y & 7)); break;
-			case BLACK:  this->buffer[tc] &= ~(1 << (y & 7)); break;
-			case INVERSE: this->buffer[tc] ^= (1 << (y & 7)); break;
+			case WHITE:  this->_OLEDbuffer[tc] |= (1 << (y & 7)); break;
+			case BLACK:  this->_OLEDbuffer[tc] &= ~(1 << (y & 7)); break;
+			case INVERSE: this->_OLEDbuffer[tc] ^= (1 << (y & 7)); break;
 		}
 }
 
@@ -384,7 +447,7 @@ void SSD1306::drawPixel(int16_t x, int16_t y, uint8_t color)
 	@param start start position
 	@param stop stop position 
 */
-void SSD1306::OLED_StartScrollRight(uint8_t start, uint8_t stop) 
+void SSD1306::OLEDStartScrollRight(uint8_t start, uint8_t stop) 
 {
 	SSD1306_command(SSD1306_RIGHT_HORIZONTAL_SCROLL);
 	SSD1306_command(0X00);
@@ -401,7 +464,7 @@ void SSD1306::OLED_StartScrollRight(uint8_t start, uint8_t stop)
 	@param start start position
 	@param stop stop position 
 */
-void SSD1306::OLED_StartScrollLeft(uint8_t start, uint8_t stop) 
+void SSD1306::OLEDStartScrollLeft(uint8_t start, uint8_t stop) 
 {
 	SSD1306_command(SSD1306_LEFT_HORIZONTAL_SCROLL);
 	SSD1306_command(0X00);
@@ -418,7 +481,7 @@ void SSD1306::OLED_StartScrollLeft(uint8_t start, uint8_t stop)
 	@param start start position
 	@param stop stop position 
 */
-void SSD1306::OLED_StartScrollDiagRight(uint8_t start, uint8_t stop) 
+void SSD1306::OLEDStartScrollDiagRight(uint8_t start, uint8_t stop) 
 {
 	SSD1306_command(SSD1306_SET_VERTICAL_SCROLL_AREA);
 	SSD1306_command(0X00);
@@ -437,7 +500,7 @@ void SSD1306::OLED_StartScrollDiagRight(uint8_t start, uint8_t stop)
 	@param start start position
 	@param stop stop position 
 */
-void SSD1306::OLED_StartScrollDiagLeft(uint8_t start, uint8_t stop) 
+void SSD1306::OLEDStartScrollDiagLeft(uint8_t start, uint8_t stop) 
 {
 	SSD1306_command(SSD1306_SET_VERTICAL_SCROLL_AREA);
 	SSD1306_command(0X00);
@@ -454,7 +517,7 @@ void SSD1306::OLED_StartScrollDiagLeft(uint8_t start, uint8_t stop)
 /*!
 	@brief  Stop scroll mode
 */
-void SSD1306::OLED_StopScroll(void) 
+void SSD1306::OLEDStopScroll(void) 
 {
 	SSD1306_command(SSD1306_DEACTIVATE_SCROLL);
 }
@@ -462,15 +525,23 @@ void SSD1306::OLED_StopScroll(void)
 /*!
 	@brief Check Connection Function
 		Check if device is on the bus asks for one byte
-	@return int16_t if less than zero = error 
+	@return int16_t if less than 1 = error 
 */
 int16_t SSD1306::CheckConnection()
 {
-	int16_t returnValue;
-	uint8_t rxData;
-	returnValue = i2c_read_timeout_us(_i2c, _OLEDAddressI2C , &rxData, 1, false, _OLEDdelayI2C);
-	if (_bSerialDebugFlag) printf("Error 1203 I2C OLED:: CheckConnection %d , %u \r\n ", returnValue , rxData);
-	if (returnValue < 1) _bIsConnected = false;
+	int16_t returnValue = 0;
+	uint8_t rxData = 0;
+	returnValue = i2c_read_timeout_us(_i2c, _OLEDAddressI2C , &rxData, 1, false, _TimeoutDelayI2C);
+	if (_bSerialDebugFlag)
+	{
+		printf("SSD1306::CheckConnection. Info \r\n");
+		printf("I2C Return value = %d , RxData =  %u \r\n", returnValue , rxData);
+	}
+	if (returnValue < 1) 
+		_bIsConnected = false;
+	else
+		_bIsConnected = true;
+
 	return returnValue;
 }
 
@@ -483,7 +554,7 @@ bool SSD1306::GetIsConnected(void)
 
 /*!
 	@brief setter for is connected status 
-	@param connected is device connected , true yes, false no
+	@param connected Set device connected , true yes, false no
 */
 void SSD1306::SetIsConnected(bool connected)
 {_bIsConnected = connected;}
@@ -501,5 +572,52 @@ bool SSD1306::GetDebugMode(void)
 */
 void SSD1306::SetDebugMode(bool debugMode)
 {_bSerialDebugFlag = debugMode;}
+
+/*!
+	@brief getter for Library version number
+	@return the version number 101 is 1.0.1
+*/
+uint16_t SSD1306::GetLibVerNum(void){return _OLEDLibVerNum;}
+
+/*!
+	@brief Gets the Number of I2C retry Attempts in event of I2C bus error
+		Set to zero for NO retry attempts. default 3
+	@return uint8_t Max number of I2C retry attempts
+ */
+uint8_t SSD1306::GetI2CRetryAttemptsNo(void){return _I2CRetryAttempts;}
+
+/*!
+	@brief Sets the Number of I2C retry Attempts in event of I2C bus error
+		Set to zero for NO retry attempts.
+	@param Attempts  Max number of I2C retry attempts
+ */
+void SSD1306::SetI2CRetryAttemptsNo(uint8_t Attempts)
+{_I2CRetryAttempts = Attempts;}
+
+/*!
+	@brief Gets delay in mS between retry attempts in event of I2C error
+	@return Delay mS
+ */
+uint16_t SSD1306::GetI2CRetryDelay(void){return _I2CRetryDelay;}
+
+/*!
+	@brief Sets delay in mS between retry attempts in event of I2C error
+	@param Delay  delay in mS between retry attempts in event of I2C error
+*/
+void SSD1306::SetI2CRetryDelay(uint16_t Delay)
+{_I2CRetryDelay = Delay;}
+
+/*!
+	@brief Gets I2C timeout value used in I2C functions calls
+	@return Delay uS, default 50000
+ */
+uint32_t SSD1306::GetI2CTimeout(void){return _TimeoutDelayI2C;}
+
+/*!
+	@brief Sets delay the I2C timeout in uS 
+	@param timeout delay in uS
+*/
+void SSD1306::SetI2CTimeout(uint32_t timeout)
+{_TimeoutDelayI2C = timeout;}
 
 // ---  EOF ---