+++ src/ssd1306/SSD1306_OLED_Print.cpp
@@ -4,10 +4,7 @@
 	@note  Port of arduino built-in print class, G Lyons 2022.
 */
 
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <math.h>
+
 #include "../include/ssd1306/SSD1306_OLED_Print.hpp"
 
 // Public Methods //////////////////////////
@@ -163,8 +160,8 @@ size_t Print::printFloat(double number, uint8_t digits)
 { 
   size_t n = 0;
   
-  if (isnan(number)) return print("nan");
-  if (isinf(number)) return print("inf");
+  if (std::isnan(number)) return print("nan");
+  if (std::isinf(number)) return print("inf");
   if (number > 4294967040.0) return print ("ovf");  // constant determined empirically
   if (number <-4294967040.0) return print ("ovf");  // constant determined empirically