+++ extra/doc/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
-* version 1.0.0 Dec 2023
+* Version 1.0.0 Dec 2023
 	* first release
+* Version 1.0.1 Jan 24
+	* Added error handling & return values to Text and 
+		Bitmap related functions.
+	* Added new I2C parameters for testing & error handling. 
+		Timeout, retry attempts amd retry delay.
+	* Wrapped the buffer pointer in a function.
+	* Added I2C test example file.